refactor: move pure functions to tableResize helper
